chore: extract output helpers
