+++ src/Console/ScanCommand.php
@@ -21,6 +21,8 @@ use Symfony\Component\Console\Output\OutputInterface;
 #[AsCommand(name: 'scan {--c|config=}', description: 'Scan the project for potential issues')]
 final class ScanCommand extends Command implements ConsoleOutput
 {
+    use OutputHelpers;
+
     protected function execute(InputInterface $input, OutputInterface $output): int
     {
         try {
@@ -76,35 +78,5 @@ final class ScanCommand extends Command implements ConsoleOutput
         (new Display($this))->report($result);
 
         return self::FAILURE;
-    }
-
-    public function newLine(int $count = 1): void
-    {
-        $this->output->newLine($count);
-    }
-
-    public function error(string|array $message): void
-    {
-        $this->output->error($message);
-    }
-
-    public function warning(string|array $message): void
-    {
-        $this->output->warning($message);
-    }
-
-    public function info(string|array $message): void
-    {
-        $this->output->info($message);
-    }
-
-    public function comment(string|array $message): void
-    {
-        $this->output->comment($message);
-    }
-
-    public function table(array $headers, array $rows): void
-    {
-        $this->output->table($headers, $rows);
     }
 }
